requirements.txt
streamlit
polars
altair
numpy
pandas
plotly-express
psycopg2